refactor(harness): share keymaxxer forge helper wire schemas
## Why

The Keymaxxer-backed GitHub and GitLab layers each defined the same
helper-process wire schemas and helpers nearly verbatim. Any change to
the shared format had to be made twice and could drift silently.

## What

- Extract shared schemas and helpers into
  `apps/harness/src/server/forge-helper-schemas.ts` (issue fields, PR
  check status / diagnostics, lifecycle, merge result, `encodeArgument` /
  repository args, `parseSerializedIssues`, `makeRequestError`).
- Decode `headPushedAt` / `createdAt` via Schema
  (`OptionalInstantFromString`: invalid strings → `null`), dropping the
  manual post-step in both layers.
- Keep GitHub-only automated-review evidence and forge-specific integer
  stdout decoders local to each layer.
- Keep intermediate schema building blocks module-local; export only
  what the layers import.
- Add a regression test that unparseable check-status instants map to
  `null`.

## Verification

- `bunx nx run harness:test` — 468 pass, 0 fail
- Biome clean on touched files

Closes #712
